Move asset directory normalizers into output_values.py
Eighth refactor pass at modules/output.py.  Two small path-normalization
helpers moved from output.py to output_values.py, joining the family
of pure stdlib-only value coercion primitives that live there.

Functions moved:
  * _normalize_asset_directory_entry  -- unescape a single Windows/UNC
    path (handles YAML-escaped backslashes, preserves UNC prefix)
  * _normalize_asset_directory_values -- normalize a list or multi-line
    string of paths

Both functions were pure (only 're' from stdlib) and only used inside
output.py.  Zero external callers.

Bonus cleanup: removed 'import re' from output.py -- no longer used.
(_parse_tmdb_person_window took its re usage to output_collections.py
in #1450; _normalize_asset_directory_entry took its re usage here in
this PR.  Nothing else in output.py needed re.)

Docstring updated: output_values.py grew a new group in the module
docstring.

Public surface: output.py re-imports both names with '# noqa: F401'.

Impact:
- modules/output.py: 2878 -> 2839 (-39 / -1.4%)
- modules/output_values.py: 203 -> 253 (+50)
- Net +11 lines total (added docstring + comment explaining the
  Windows/UNC edge cases)

Cumulative on output.py: 3833 -> 2839 (-994 / -25.9%)

All 1077 tests pass. Ruff + black clean.
